Boss stage 3 done

+++ Knowledge_Is_Power/BossSmallBullet.java
@@ -8,7 +8,7 @@ import greenfoot.*;  // (World, Actor, GreenfootImage, Greenfoot and MouseInfo)
  */
 public class BossSmallBullet extends EnermyBullet
 {
-    public BossSmallBullet(int s,int r,String img_name){
+    public BossSmallBullet(double s,int r,String img_name){
         super(r,20,20,5);        //default size 20*20, not through, 5 damage
         move_speed = s;
         
@@ -17,7 +17,7 @@ public class BossSmallBullet extends EnermyBullet
         setImage(image);
     }
     
-    public BossSmallBullet(int s,int x, int y, String img_name){
+    public BossSmallBullet(double s,int x, int y, String img_name){
         super(0,20,20,5); //default size 20*20, not through, 5 damage
         move_speed = s;
         turnTowards(x,y);
